Remove CachePolicy and WDTOptions
